requirements.txt
fastapi
uvicorn[standard]
telethon
pyrogram
tgcrypto
python-multipart
python-dateutil